Added RenderBundleResourceContextProps to render detail pages
